refactor: extract program microcycle to views/program/ProgramMicrocycle.jsx
Part of #77 (PR 2 of 3 — the program decomposition). Splits the "2-Wk Cycle"
sub-tab out of ProgramView.jsx into its own component, following PR1 (#110).

Also fixes a live bug found while rewriting ProgramView's imports: the phases
tab referenced `C[s.type]` without importing `C` from constants/ui.js, which
would throw a ReferenceError on render.
